update files structure, separete services.py to 3 servises: user, session, todo
