+++ .gitignore
@@ -1,5 +1,6 @@
 __pycache__
 .venv
+.env
 .vscode
 db_data
 app.log
